requirements.txt
pywebview==4.4.1
psutil==5.9.6
requests==2.31.0